feat: :sparkles: Redesign all core functionality
Change file structure to be more intuitive and sensible. Add classes for session and user, remove code made redundant by this addition. Restructure project to install properly as a package and run in CLI, __main__ is used as entry point. Add setup.py for package installation and install.sh script to automate for devs. Update .gitignore